PS interpreter - fix cvrs with 64-bit integers
Bug #702195 "Integer overflow in cvrs on Windows"

Another 64-bit integer fix, we were casting the result of the radix
calculation through a long which on Windows is not 64-bit (on Linux it
is), leading to truncation o the result.

Fixed by casting through ps_uint instead. We also need to check for
CPSI mode here as we must use a 32-bit intermediate integer if we are in
CPSI mode and emit a rangecheck if the integer overflows 32 bits.