Don't set base space for certain pattern types
Use pattern_instance_uses_base_space() (like the gs/pdf implementation
does) to check whether the particular type of pattern expects a base
space.  Specifically, PatternType=1, PaintType=2 uses the base space,
but PaintType=1 does not.

This causes a progression in a bunch of samples that were
incorrectly trying to read the colors from the stack when they weren't
there.

See tests_private/pdf/sumatra/692135_-_segfaults_when_tiling.pdf
Note that this sample still has a bug, my trimmed 692135-trim.pdf
renders small box as empty instead of with lines in it.

Also tweaked the error handling to not continue processing a pattern
that failed to be set.  This causes a progression in the file
tests_private/comparefiles/Bug692783.pdf which has missing patterns
(i.e. they are referenced but don't exist).  At this point gpdf now
renders this like Adobe does, while the gs/pdf version is missing some
words in the background on pages 2 and 3.