Add support for 64bit and size_t paramaters.
Cope with both int64_t's and size_t's with appropriate
rangechecks when casting will cause problems.

-d values on the command line are read as int64_t's.

If we ever find outselves wanting to assign a size_t where
the 63rd bit matters, then we may have to rethink this :)

We also change various params (such as MaxBitmap, MaxPatternBitmap
etc) to be size_t's. This affects the device structure itself.