pdfwrite - preserve ToUnicode CMaps with non-standard mappings
Bug #702201 "pdfwrite changes ligature"

The title is incorrect, pdfwrite does not 'change ligature'. The PDF
file uses an ffi ligature glyph from the embedded font, but the
ToUnicode CMap does not map that to U+FB03, it maps it to 3 Unicode
code points; f, f and i.

When we emit a simple font we check its Encoding to see if it needs a
ToUnicode CMap, if it doesn't then we don't emit one. Because all the
glyphs in this font are named, and have standard names, we don't need to
emit a ToUnicode CMap and so we don't (keeps the file size down).

However, that means that PDF consumers will map the ligature glyph to a
ligature ffi when copy/pasting (it has no other effect).

Here we check the Unicode code points as they are added to the internal
record. If any of them is not 2 bytes long, then its not a standard
mapping (simple fonts can only map to a single point). In that case we
note the exception and when we come to write the font, in the code that
checks the Encoding to see if we need to write a ToUnicode, we check
this value, if we've seen a multiple code point in the ToUnicode CMap
then we don't elide it, we embed the one we picked up.

This is not actually a bug IMO, there's nothing wrong with the mapping
as it stands because an ffi glyph in the document really should map to
a single Unicode code point ffi ligature. However, the creator of the
PDF file obviously didn't want that, so we;ll now honour it.

NB this only affects the emission of a ToUnicode CMap, which is only
used for copy/paste/search. The PDF contents are not affected by this
commit.