Bug 702191. Change Tr modes 2 and 6 to use fill_stroke operation.
Thanks to Ken Sharp who made the change to the PDF interpreter pdf_ops.ps.
This showed a few files that had SEGV. This was due to both the fill and
stroke having pattern colors where the pattern bits did not fit in the cache.
The stroke color could eject the fill pattern from the cache. The fix is to
lock the fill pattern in the cache as well as the stroke pattern (and unlock
it as well after the fill_stroke).