Fix bug 702142: SMask profile 'swap' could cause clist to use wrong profile
The SMask 'swap' of profiles for the 'default' colorspaces in the icc_manager
was not being detected by the clist logic since the colorspace id was still
set to the cs_Device*_id value (1, 3, or 4).

The fix is to have the clist writing check for unknown color_space check BOTH
the color_space.id and the color_space.space, and flag 'unknown' if either
differed. The color_space.space includes the icc_profile info.

This results in a few progressions as well as fixing the file of this bug.