Repair device subclassing after fill+stroke commit
Bug 702221 "Ghostscript 9.51 - SIGFPE when -dFirstPage or -dLastPage is set"

The fill+stroke commit had an error with the default method for handling
fill_stroke_path which passed the subclassing device as a parameter to
the child, instead of passing the child device. This caused seg faults,
floating point exceptions, hangs, and possibly other problems.

In addition the object filtering device had been modified to pass the
fill+stroke method to the default handler, instead of checking to see
if vector operations should be dropped.

This commit fixes both problems.