Fix pdfwrite with TJ
The pdfwrite device, somewhat bizarrely, doesn't use the graphics
state current point to detect a shift in text x co-ordinate. It instead
uses the last point in the path.

For PostScript or PDF input at least, this makes no appreciable
difference as there's no way to move the current point without using
some form of move operator, which adds a point to the current path.
However for pdfi when executing the TJ operator we were just updating
currentpoint. This was causing pdfwrite output to be incorrect, because
it didn't look at the current point and therefore was using the wrong
starting point.

Fix that here by using rmoveto instead of calculating the distance
transformed by the CTM and updating currentpoint manually.

Credit where its due; Nancy spotted this and did most of the work on it.