Bug 702115: use the offical method to detect gcc
In the early days of autoconf there was no defined way to check if the selected
compiler was gcc (or compatible) or not. Projects used the autoconf internal
shell variable ac_cv_prog_gcc to get that info.

autoconf introduced the GCC shell variable as the official way to know (a long
time ago), and we never revised our configure.ac to reflect that.

This does so.