Partially implemented Stamp annotations with no AP
This draws the frame and has all the data for the different Stamp types,
but doesn't do the text.

Will implement text later.

The only test file I found on the cluster that uses Stamp annotations
is tests_private/pdf/PDF_1.7_FTS/fts_32_3204.pdf.  That file can be
trivially edited to test all the various types if desired.