Fix type 3 fonts with pdfwrite
We were supplying a function to return a glyph name givven a glyph index
which used the font's Encoding. While efficient, this failed if the
function was called after the font had been released.

It turns out that pdfwrite expects that the name will still be available
even if the font has been released. Normally this works because the name
that is returned uses the name table, which is persistent to the end of
job.

So modify the 'pdfi_glyph_name' function to retrieve the glyph name
from the Encoding, add it to the name and then use that index to
return the name and length. The pdfwrite device stores a pointer to the
name, so we don't need to worry about matching the glyph index to the
name later.

There is a problem with the existing pdfi_name_from index() which was
being used to retrieve separation names for halftones, it won't work
as currently configured because there appears to be no way to get back
to the interpreter context from the point we are at when the function
is executed.

Fixing that is going to be more of a problem, for now I've moved it to
its own routine and changed pdfi_name_from_index() to take the context
as a parameter.

The loop finding a name didn't terminate properly and always returned
an undefined error, fix that by breaking out of the loop when we find
a name which matches a requested index and returning 0.