Bug 702190: Set CFLAGS and CFLAGSAUX separately in Makefile
Building with -DGS_NO_FILESYSTEM allows support for environment without a
proper file system, by stubbing out the gp_ calls that would normally access
the file system.

Obviously, for Ghostscript to function, that means it requires its resource
and initialisation files to be built into the binary in the ROMFS.

A problem arises if you include the -DGS_NO_FILESYSTEM CFLAG when building the
mkromfs utility, which creates the ROMFS, since denying it access to the file
system results in an empty ROMFS. Previously, setting CFLAGS to include
-DGS_NO_FILESYSTEM when calling configure resulted in that flag being applied to
the main build, and the auxiliary tools build (including mkromfs), unless the
build specified was a cross-compilation.

This splits the CFLAGS and CFLAGSAUX so they are set separately in the Makefile
so configure can filter out unwanted flags from the auxtools flags.

Currently, the only one being filtered is -DGS_NO_FILESYSTEM