Ensure dictionaries and arrays are of correct type
When we dereference an object, and its a free object, we return a null
object instead. This can cause seg faults if we then try to get an
object from it in the belief its a dictionary or array. Possibly we
should not return a null object, but for now add type checkign to ensure
this can't happen