Further fix for text processing
Although we had added the glyph names to the name table, we were still
freeing the name table too early. We freed the name table at the end of
each input file, but we don't write the Encoding until we close the
pdfwrite device, and we don't do that until the interpreter is closed.

So defer freeing the name table until the interpreter exits.

This causes a number of text progressions with pdfwrite and ps2write
and will hopefully resolve a number of indeterminisms.