The /Prev xref from an XRefStm need not itself be an XRefStm, it could be a regular xref. So when handling a /Prev in this case check what we read from teh file, if its the keyword 'xref' then handle it as a normal xref, otherwise assume its an XRefStm.
When dealing with a hybrid file, after reading the XRefStm we still
need to check any /Prev xref, and its possible (for an edited file) that
the XRefStm and the xerf will both reference the same /Prev. To prevent
this causing a problem, mark and clear the circular reference checking
around the XRefStm reading.