PDF interpreter - cope with Form XObjects leaving junk on the stack
No bug report, customer supplied file. The file has a Form XObject
where the stream ends:

/Image20 Do Q
 EMC  Tendstream

which is clearly invalid. This leaves a T lying around which gets
converted to a name and remains on the stack. When we come to complete
the form processing we expect that the stack will be clean, because it
isn't we end up corrupting an internal variable, which causes the
*next* Form XObject to fail.

We cope with this here by adding code to check the objects on the stack
and remove them one by one until we get back to a dictionary which
contains a name we expect to find (because we pushed it on entry to the
procedure) or run out of stack objects in which case we will throw an
error.