Squashed commit of the opacity_shape branch
This commit removes shape.alpha and opacity.alpha from the
graphic state.  This involved removal of methods in the
postscript based PDF interpreter, changes in the pdf14 device,
changes in pdfwrite, the XPS interpreter, and the PS
transparency test file that is present in examples and on
the cluster.  The goal of  this commit was to avoid
confusion and conflicts between shape.alpha opacity.alpha and
the stroke and fill alpha values (CA ca).  This change
resulted in progressions in the XPS file.  There were changes
in a couple PDF files that required some investigation.  In
particular, the change in the dashed lines of Bug694981.pdf.  Robin
and I looked at this.  Robin determined it was due to
differences in stroke_add and stroke_fill not giving identical results.
In the end, the new output looks closer the what AR provides.