Redefine deprecated .setopacityalpha and .setshapealpha
.setopacityalpha and .setshapealpha have been replaced with
.setfillconstantalpha/.setstrokeconstantalpha and .setalphaisshape, operators
which actually do what PDF requires.

As an interim measure, to allow users that rely on these operators some time
to adjust, define versions of .setopacityalpha and .setshapealpha which
approximate what the old implementations did.

As with the rest of the transparency operators they only exist when gs is
given the -dALLOWPSTRANSPARENCY option. Also, they are in pdf_main.ps just so
they grouped in proximity to where the rest of the ALLOWPSTRANSPARENCY logic is
implemented.