Change MaxLocalVM and MaxGlobalVM to be size_t instead of 'long'
The sizeof long differs between 64-bit Windows and linux, and
Robin changed the Max values for other params to be size_t in
commit d7961033. This change extends that to the other memory
size related params.

Also fix vm_threshold to be int64_t, not size_t since that
corresponds to the parameter setting in zusparam.c. This is
also limited to the max signed value that corresponds to
size_t (MAX_VM_THRESHOLD) in case size_t is < 64 bits.