Fix handling of an error in a stream
If the stream gets an ioerror, now treat that as an EOF so we can recover
sensibly.  Flag this with a warning flag.

This fixes file tests_private/pdf/sumatra/691958_-_jbig2dec_crash.pdf
which was exiting with an error -12 (ioerror) and now will render a
blank page.

The test file has a totally bogus JBIG2Decode filter in it.