PDF interpreter - Fix incorrect CIDFontType2 with Type1C descendants
Customer bug report. The file contains a CIDFont declared like this:

16 0 obj
<<
...
  /Subtype /CIDFontType2
...
>>

The /Subtype /CIDFontType2 means that the font is a CIDFont with
TrueType outlines. However the actual font:

10 0 obj
<<
  /Subtype /Type1C
  /Length 49432
  /Length1 49432
>>

is a Type1C (Compact Font Format) font. The PDF file is, quite simply,
incorrect. The Subtype should be /CIDFontType0. Interestingly we already
have code in the CIDFontType2 case to handle an incorrect value of
Subtype being set to /CIDFontType0 so this is the reverse case.

If we discover we are reading a Type1C font, and the parent is a CIDFont
with a Subtype of CIDFontType2, then change the parent dictionary to
have a Subtype of /CIDFOntType0C and read the font as a CIDFont.