Don't clear stacks at end of stream if there were no errors
Its potentially possible to have something like:

stream
....
0 0
endstream

stream
m
....
endstream

The actual example has a TJ where the argument is in the stream
preceding the stream containing the TJ.

By clearing the stack between streams we were discarding the arguments
from the stack. This commit removes that (we retain stack clearing in
the case of something going wrong) but clears the stack at the end of
each page instead (Nothing should persist between pages).