Move transparency and spot colour checking to per page from doc level
For spot colour capable devices we need to set the number of spots
on a per-page basis, rather than per instance of the device.

Previously we were checking for transparency and spot colours, and
configuring the device with the number of spots, when the PDF file was
opened, which caused an odd seg fault.

Testing indicates that checking each page as it is processed is
certainly no slower, and may actually be faster, than the existing
scheme, somewhat to my surprise. I suspect this may be because we cache
the colour spaces etc when checking them, and they are then still in the
cache when the page is processed. In the 'up-front' scheme we win if
the colour spaces are reused between pages but it seems like this is
not a common situation.

This solves the seg fault, a few files show some pixel level differences
with the psdcmyk device and transparency_design_guide shows a slight
(and somewhat baffling) progression with the psdcmyk device.