Fix memory leak in error path of icc color stuff
Sample: image_2bpp.pdf

Fixed a definite leak in pdfi_create_icc() on error path.  Wasn't
decrementing pcs.

Cleaned up some related error handling in pdfi_create_iccbased() that
was probably okay, but it was a mix of returning in the middle of the
code vs. jumping to the cleanup at the end.