Bug 702322: fix uninitalized data reads
gs_scan_token() keeps a local copy of the scanner state, and copies from and to
the scanner state that is passed into it. There are several code paths that
can leave some important entries in the structure uninitalized when we copy the
local copy to the parameter.

This just ensures those specific entries are always set to *something* sane.