Add/modify code to record the 'indirect' object number and generation
This is for future support of decryption code. When decrypting data
we must create a decryption key which is partially based on the object
and generation numbers of the PDF object.

For streams this isn't a problem because streams are always indirect
references. However we also need to decrypt strings, and strings may
be defined directly (eg as entries in a dictionary). In this case the
specification says we must use the object number of the enclosing
indirect object. This is complicated, because dictionaries and arrays
can themselves be defined directly, in which case we need the object
and generation numbers of the object which encloses the composite
object. And so on....

Here we add two new entries to all objects, the indirect_num and
indirect_gen. In order to keep the storage down, we redefine the
existing object number as a 32-bit value instead of a 64-bit (its still
larger than Acrobat permits, which is realistically all that matters).
We also define the indirect_gen as a 16-bit number, because we only
need the bottom 2 bytes for decryption.

This does of course spread throughout the code a little as we need to
pass around, or copy, the numbers from one object to another.

Also fix some minor problems with revision 4 decryption (which isn't
actually complete yet anyway).