Fix memory leak related to closing filters
If there's a chain of filters, and one of them reaches EOF (possibly
because of an error in the data) then all the filters weren't getting closed,
causing a memory leak.

I fixed this by setting the "close_at_eod" flag to false when we open
our filters.  There might be a different way to fix this, but since I
think we always explictly close our filters anyway, I think this is
fine?  The problem was that the file_close_file() being called didn't
actually close the other filters in the chain, but it nulled out the
strm pointer so pdfi_close_file() couldn't find the rest of
them. Something like that....??

File was page 1 of image_2bpp.pdf (my trimmed file image-trim1c.pdf)

Also cleaned up some image error handling that I noticed along the
way, but not directly related to this issue.