Fix ICCBased spaces with setcolorspace
When we create colour spaces we take a pointer to return the created
space in (for the benfit of, eg, images). If that pointer is NULL
its a signal that we don't want the space back, we just want it set.

In general the creation routines create a local space and the last
step before returning is to see if the pointer is NULL and set it to
 the created space if it is not.

The ICCBased code, however, wasn't doing that. It was using the passed
in pointer immediately to create the colour space. If the pointer is
NULL then this causes a seg fault.

Change the ICCbased colour creation to follow the other spaces.