Implement revision 4 decryption
There are a number of changes in here; the applicatin of the decryption
filter has been moved from pdfi_interpret_content_stream() to
pdfi_filter() so that it affects all cases where filters are applied,
not just content streams.

Fixed a couple of missing pdfi_countdown cases at the same time.

There is a new pdfi_filter_no_decryption for the benefit of the xref
parsing, because xref streams are *not* (can't be) encrypted and must
be read without a decryption filter.

Fixed a typo where the object and generation numbers were reversed in
xref parsing at the same time.

Increased the 'min_size' of the Arc4, AES and Subfile decode filters to
2048 because setting these too small causes the stream code to hang.

Changed the test of EncryptMetadata when applying the R4String (whcih is
4 bytes of 0xFF) as it was inverted.

Removed the (empty) revision 4-specific password tests as these are not
needed.

Implemented the 'sAlT' step of object key computation for revision 4
using the AES filter (and for the future revision 5)

Read the StmF, StrF and Cypt Filter entries from the Encrypt dictionary
for use by revision 4 decryption (choice of RC4 or AES decryption).