PDF interpreter - reject insane Tc values early
No bug report, the file is from a customer and confidential.

The file has been created by a bfo.com product, but then additionally
modified by an unknown application. The problem is that the file
contains this sequence:

1 0 0 0.999999 21.6 9.900086 cm
/R3 9 Tf -2147483647 Tc -0.39301 Tw
4.8 -21.95 Td
[( )]TJ

The parameter supplied to Tc is clearly mad. Not invalid but it attempts
to move the current point (after drawing the space) by 2,147,483,647
unscaled text units. This works out at around 471 miles. Given that this
value is the maximum negative  32-bit integer (0xFFFFFFFF) I doubt this
is actually intentional.

The scaling by the CTM causes a limitcheck error when trying to draw
the text. We could fix this by running the TJ in a stopped context
but that would be expensive performance-wise because TJ is a very common
operator.

Instead this commit checks that the result of Tc, when scaled by the
CTM, will fit into our fixed-point representation. If it does not then
we throw an error there and then. The PDF interpreter then catches that
during the Tc operator (which is less commonly used than TJ) and emits
a warning, while ignoring the Tc value.

This may, of course, result in incorrect output, though it seems
unlikely given the size of the numbers involved, which is why we give
the user a warning.