Fix to ignore Length 0 encrypted streams
The code was trying to decrypt them, resulting in garbage.  This fix
just treats them like a regular stream if Length is 0.

Maybe there's a better fix, but this was a "if it hurts, don't do
that!" approach...

Test File: tests_private/pdf/sumatra/x_-_missing_annotations_on_page_1.pd