Add code to handle revision 5 & 6 decryption
Add sha2.c to the build for revision 5 & 6 decryption

Changes to length of O and U, these can now be 48 bytes. For revision
2-4 we only use the first 32 bytes.

Permit O and U to be greater than the specified size (32 or 48 bytes)
and only use the first 32 or 48 bytes.

Some reference counting fixes.

Still TODO : string decryption, import the locale_to_utf8 code for
Linux and Windows, and attempt to turn passwords into UTF8 if they
don't work as ASCII. Tidying up, SHA256 for revision 5 would be nicer
as revision 6 (less messing with streams). Check for memory leaks and
reference counting problems.