Fix tricky issue
This had to do with a clist case where there was an empty
group within a band and there was no page group and the
only drawing occurred after the group pop.   This caused
significant confusion that had to be addressed by making
the test for creating a buffer for the context a little
more nuanced.