Decrypt strings, as required, from encrypted PDF files
This turns out to be complicated. Unlike streams, which are always
indirect references and cannot be defined inside other objects, strings
can be indirect references or directly defined, and can be created
in the course of a content stream as well as for array and dictionary
entries.

In addition, directly defined strings have no object number, and so must
use the object number and generation of whichever indirectly referenced
object contains them.

The inheritance of object numbers was handled in an earlier commit, we
simply use the inherited values now if required.

We have a new boolean in the PDF context, which I was hoping to avoid
doing but it seems unavoidable and in any case makes the code much
simpler. When decrypt_strings is true the parser will call
pdfi_decrypt_string() for any strings which it creates.

This flag is generally true, but it is false whenever we are running a
content stream (because content streams are already encrypted and we
don't additionally encrypt strings in that case). Note that this flag
will be false when running Form XObjects as well as Page Contents,
because its possible (ag Annotations) to execute a Form outside a Page
Content stream.

Objects which are defined as indirect references (eg Colour spaces) can
be used, and dereferenced from inside a content stream, so we set the
flag to true for the duration of a dereference operation, preserving
the state around the operation.

The actual decryption code checks the object number (of the string or
the enclosing indirect object) to see if it is defined in an ObjStm;
Like content streams ObjStms are streams and therefore already encrypted
again we don't additionally encrypt strings in an ObjStm, so we need to
avoid decrypting these kinds of objects.