Fixed memory leak when an object wasn't at the expected offset
There was a complicated mess caused by the case where an object that
was located at a certain offset was not the correct one.

File: tests_private/pdf/sumatra/incompletely_repaired.pdf

In this file, object 1 is completely missing, though the xref says it
should be at offset 0.  Object 2 is at offset 0.

We tried to load this object several times, and each time got an
error, but left various allocation issues.  It would put it in the
cache each time it looked it up, but it was putting it in the cache as
object 2, so there were like 3 object 2's in the cache. Other weird
stuff.

I decided the easiest thing was to add 'expected_num' to
pdfi_read_object() so it can detect the wrong object early, before it
does a lot of extra work, puts it in the cache, etc.  If the object
number isn't known (it wasn't obvious to me on the
pdfi_read_object_of_type() path, but maybe it is also known there?)
then just specify a 0 and it won't check.