Fixed some error handling for broken files
Added some error messages for better reporting, but also two
significant changes:

1) In pdfi_page_render(): do the pl_finish_page if not pdfstoponerror.
This can cause pages that had errors to be partially rendered (or at
least a blank page).  Previously some multi-page jobs with errors
would not print all the pages.  For instance a job with 86 pages only
printed 84 of them.

2) In pdfi_process_file() in main loop over the pages:
Clear the error code if not stopping on error.  Single page jobs (or
jobs with error on last page) were causing an exit with an error code
even when pdfstoponerror not set because the code was still negative.

The end result of these changes should be fewer interpreter exits with
error codes and more blank pages rendered, which will more closely
match what gs does.