Fixed bug in parsing XRef with 0 values in /W
File: tests_private/pdf/sumatra/1146_-_repairing_ignores_encryption_regression_from_r2057.pdf sumatra_1146.pdf

This was building an XRef where all entries were marked free.
It has an XRef with entries of /W [0 2 0]

Spec says 0 width field for the first value (type) should default to
1, but it was defaulting to 0 (which meant free).

Also the logic for figuring out what the field_width is, so how many
bytes to allocate, was kind of obscure.  I simplified it and made a new
variable instead of overloading 'j', for clarity.

Changed the code to not try to read anything if the width of the field
is 0, just use the default.