Fix Bug 702014: ps2epsi problems (regression?).
The ps2epsi and ps2epsi.ps were quite a mess, and relied on non-standard
operators only available with -dNOSAFER since we made SAFER the default.
Even running ps2epsi with -dNOSAFER (by export GS_OPTIONS=-dNOSAFER)
had problems -- it could not even convert examples/colorcir.ps to EPSI
(throwing an error about colorcirsave). Both of the aforementioned
files are replaced with this commit, and the ps2epsi.bat and ps2epsi.cmd
are updated to use the new ps2epsi.ps file

Also, input files with multiple pages were never correct since the preview
was derived only from the first page. The ps2epsi.ps checks if the file is
already an EPSF, then don't use eps2write on the input, but warn if the
%%Pages comment indicates that there is more than a single page.

Note that -dNOOUTERSAVE is needed in the invocation so that if we use
eps2write, the 'restore' will close the device and write the tempfile
with the EPSF.

Then ps2epsi.ps reads the input file, capturing the BBox, stops after
%%EndComments, removes any existing preview, constructs a new preview
using the 'bit' device with appropriate parameters, writes the preview
in the EPSI Preview format (using writehexstring) and then copies the
remainder of the EPSF file to the output.

The ps2epsi.ps has some checking for valid invocation, since someone
may try to use it without the scripts.

Since all of the actual conversion is done by the ps2epsi.ps, this change
makes windows and linux produce equivalent files.