Alter the handling of errors with /Mask in image dictionaries.
Not finding a /Mask isn't an error, it just means hte image doesn't have
one. However e test file openjpeg_166_-_buffer_overflow.pdf shows
that the current Ghostscript implementation ignores any error when
finding the /Mask.

We should not be relying on an 'undefined' error as the only one to be
ignored. Add code to record the warning, and only actually exit if
PDFSTOPONWARNING is true.