When reading a compressed ObjStm, check that the ObjStm contains the object number of our target object, at the index we expect to find it. Throw an error if it does not.
Also do some tidying up. We were already checking the object created by
pdfi_read_object to see if it was the one we wanted, so we can do away
with passing the expected number into pdfi_read_object.

We also don't really need pdfi_read_object_of_type, since we only call
it in one place. Just add the type checking to the caller and do away
with the function. In the same location, check the object number of the
returned object and

Re-arrange the code so that pdfi-read_object precedes its caller.