Move the caching responsibility from pdfi_read_object to the caller.
Add some checks to pdfi_read_object, don't throw an error if we have an
endstream but a missing endobj (but note it), unless PDFSTOPONERROR is
true.

Rename skip_white to pdfi_skip_white, since it is not static.

If we find a spurious '>' then don't consume the next byte of input if
it is not another '>' (to make a dictionary mark).

When repairing a file and scannning for objects, start each scan by
skipping white space (we were off-by-one for offsets before).

Store the offset of each of the last 3 objects while repairing. This is
so that if we find a 'obj' while looking for an endobj we can determine
the location of the object number, which gives us the offset for the
start of the new object.

This resolves a few errors while repairing files.